🎨  App restructure - Folder & file names with Docs update.
