+++ app/notification/notification_operations/template_types.py
@@ -1,10 +1,10 @@
-from app.notification.custom_exceptions import NotificationError
+from app.notification.notification_operations.custom_exceptions import NotificationError # noqa
 
 
 def email_recipient(data: dict, example_data: dict, notification_class):
     """
     Function matches with the correct template type &
-    calls the relevant function from process_notification_data.py.
+    calls the relevant function from process_data.py.
     """
     match data.get("type"):
         case "MAGIC_LINK":
